backend/requirements.txt
fastapi==0.109.2
fitz==0.0.1.dev2
Pillow==11.2.1
pytesseract==0.3.13
python-dotenv==1.0.1
SQLAlchemy==2.0.27
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
pydantic-settings==2.1.0
google-generativeai==0.3.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
